requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
openai==1.35.13
Pillow==10.3.0
requests==2.32.3
numpy==1.26.4
Jinja2==3.1.4